requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.1
werkzeug==3.0.3
email-validator==2.1.1
pytest==8.2.2
pytest-cov==5.0.0
